+++ app/Http/Controllers/SaidaController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\Saida;
+use App\Models\Cliente;
+use App\Models\Produto;
 use Illuminate\Http\Request;
 
 class SaidaController extends Controller
@@ -16,11 +18,38 @@ class SaidaController extends Controller
      }
 
    public function store(Request $request) {
+        
+        $cliente = Cliente::find($request->id_cliente);
+        if ($cliente == null) {
+            return response()->json(['erro' => 'Cliente não encontrado']);
+        }
+
+        
+        $produto = Produto::find($request->id_produto);
+        if ($produto == null) {
+            return response()->json(['erro' => 'Produto não encontrado']);
+        }
+
+        
+        if ($cliente->idade < $produto->faixa_etaria_minima) {
+            return response()->json(['erro' => 'Cliente não possui idade para comprar este produto']);
+        }
+
+        
+        if ($produto->quantidade_estoque < $request->quantidade) {
+            return response()->json(['erro' => 'Quantidade insuficiente em estoque']);
+        }
+
+        
         $saida = Saida::create([
             'id_produto' => $request->id_produto,
             'id_cliente' => $request->id_cliente,
             'quantidade' => $request->quantidade,
         ]);
+
+   
+        $produto->quantidade_estoque = $produto->quantidade_estoque - $request->quantidade;
+        $produto->update();
 
         return response()->json($saida);
     }
@@ -33,6 +62,14 @@ class SaidaController extends Controller
         if ($saida == null) {
             return response()->json(['erro' => 'Tarefa não encontrada']);
         }
+
+        $produto = Produto::find($saida->id_produto);
+        if ($produto == null) {
+            return response()->json(['erro' => 'Produto não encontrado']);
+        }
+        
+        $produto->quantidade_estoque = $produto->quantidade_estoque + $saida->quantidade;
+        $produto->update();
 
         $saida->delete();
